requirements.txt
Box2D==2.3.10
certifi==2020.12.5
cffi==1.14.5
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
evdev==1.4.0
future==0.18.2
glfw==2.1.0
gym==0.18.0
imageio==2.9.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==7.2.0
pycparser==2.20
pygame==2.0.1
pyglet==1.5.0
pynput==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-xlib==0.29
scipy==1.6.3
six==1.15.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
